+++ packages/typera-express/src/middleware.ts
@@ -7,6 +7,12 @@ export type Middleware<
   Response extends commonResponse.Generic
 > = commonMiddleware.Middleware<RequestBase, Result, Response>
 
+export type ChainedMiddleware<
+  Request,
+  Result,
+  Response extends commonResponse.Generic
+> = commonMiddleware.Middleware<Request, Result, Response>
+
 export type Generic = commonMiddleware.Middleware<
   RequestBase,
   any,